refactor(cache_simulator): split algorithms into separate .c files and use switch-case in demo
